requirements.txt
mysql-connector-python==8.0.33
pip install pandas openpyxl